refactor: WIP: separate compile from language feature
code changes should drive the compilation. Therefore, compile the code
in the didSave, didChange events and update the global. The language
features the just get the information from the global.

This does currently not work because I wrongly used an IO Global as
state. Instead it should be a Global only that we can get with
performUnsafe. Otherwise we stay in IO and due to the laziness no work,
that means no compilation is done.

Search for the marked TODO in the code to see where to make the changes.

Rebase on main first!