Split Compile from Language Feature (#33)
* WIP: how to make the frege compiler configurable

explore the effort of reading a gradle plugin property with
the gradle tooling api.

* feat: add shouldCorrectlyConfigureExtraClasspath test

Expose two environment variables to configure the frege compiler:
- FREGE_LSP_SOURCE_DIR
- FREGE_LSP_EXTRA_CLASSPATH

* refactor: split compilerHelper into compileOptions, compileGlobal and
compileExecutor modules

* refactor: WIP: separate compile from language feature

code changes should drive the compilation. Therefore, compile the code
in the didSave, didChange events and update the global. The language
features the just get the information from the global.

This does currently not work because I wrongly used an IO Global as
state. Instead it should be a Global only that we can get with
performUnsafe. Otherwise we stay in IO and due to the laziness no work,
that means no compilation is done.

Search for the marked TODO in the code to see where to make the changes.

Rebase on main first!

* refactor: use Global instead of IO Global as state

Resolves all the todos introduced by the previous commit.

* chore: update version