feat: compile frege files (#35)
Local frege imports are resolved with .class files on the classpath.
For this we need to compile the frege files to java first and then run
the java compiler to compile to .class files.

However, since make mode is not yet completely implemented, the imports
are not resolved interactively. As a result, you may have to restart the
language server completely if you change the dependency of a module.