feat!: change to incremental document sync
Breaking change: make mode needs to read from disk to resolve dependencies
and recognise the project structure. As a result, I switched the single
source of truth of the document state to disk altogether whereas before it
was in the client's buffer.

This results in the following lifecycle change:
Instead of getting the document content through the didChange
notification (ownership in the client's buffer), I read the content in
the didSave notification from disk.

As a consequence, frege files are only compiled when saved and language
features may show outdated info until the files are saved. However, this
can easily be automated using autosave in the client.