Make Mode Support (#37)
* WIP: make mode exploration

The make mode seems to work correctly. However I am unable to extract
the global of the compiled module yet. In addition, I cannot yet handle
the asynchronous nature in a test case.

* refactor: move all make mode explorations into its own module

Add one new test and ensure that all tests pass.

* feat: make mode now finally works

* refactor: adapt make mode to architecture

The lsp4j `FregeTextDocumentService` class should only call into the
service layer and not directly into Frege core classes like
`CompileMakeMode`. As a result, I added a `CompileService` class which
takes care of the compilation.

The compilation's call flow is now equal to the ones of the language
features: FregeTextDocumentService -> (Compile)Service ->
(CompileExecutor)LSP -> CompileMake (Frege core class).

Additional things done:
- various cleanup
- various unifications

* fix: create compile output dir on initialize

The output directory must exist so that the class loader can correctly
resolve dependencies and imports on the classpath.

* feat!: change to incremental document sync

Breaking change: make mode needs to read from disk to resolve dependencies
and recognise the project structure. As a result, I switched the single
source of truth of the document state to disk altogether whereas before it
was in the client's buffer.

This results in the following lifecycle change:
Instead of getting the document content through the didChange
notification (ownership in the client's buffer), I read the content in
the didSave notification from disk.

As a consequence, frege files are only compiled when saved and language
features may show outdated info until the files are saved. However, this
can easily be automated using autosave in the client.

* fix: change hardcoded local paths to relative to make tests pass